+++ app/src/main/java/com/example/covid/LeftFragment.kt
@@ -1,0 +1,22 @@
+package com.example.covid
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+
+/**
+ * A simple [Fragment] subclass.
+ */
+class LeftFragment : Fragment() {
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_left, container, false)
+    }
+
+}
