+++ app/src/main/java/com/example/covid/MainActivity.kt
@@ -4,8 +4,11 @@ import android.content.Context
 import android.content.Intent
 import android.net.ConnectivityManager
 import android.os.Bundle
+import android.view.MenuItem
 import androidx.appcompat.app.AlertDialog
 import androidx.appcompat.app.AppCompatActivity
+import androidx.fragment.app.Fragment
+import com.google.android.material.bottomnavigation.BottomNavigationView
 import com.google.gson.FieldNamingPolicy
 import com.google.gson.GsonBuilder
 import kotlinx.android.synthetic.main.activity_main.*
@@ -19,82 +22,35 @@ import okhttp3.Request
 
 class MainActivity : AppCompatActivity() {
 
-
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
-        if (isNetworkAvailable() == false) {
-            alertForInternet()
-        }
 
-        centerImage.setOnClickListener {
-            content.startRippleAnimation()
+        main_nav.selectedItemId = R.id.middle_menu
+        val ft = supportFragmentManager.beginTransaction()
+        ft.replace(R.id.main_frame, MiddleFragment())
+        ft.commit()
 
-            fetchData()
-
-        }
-    }
-
-    private fun fetchData() {
-        GlobalScope.launch {
-            val okHttpClient= OkHttpClient()
-            val request= Request.Builder().url("https://api.covid19india.org/data.json").build()
-            val response = withContext(Dispatchers.IO){okHttpClient.newCall(request).execute()}
-            if (response.isSuccessful) {
-                val gsonBuilder = GsonBuilder()
-                gsonBuilder.setFieldNamingPolicy(FieldNamingPolicy.LOWER_CASE_WITH_UNDERSCORES)
-                val gson = gsonBuilder.create()
-                val data = gson.fromJson(response.body?.string(), TotalResponse::class.java)
-
-                launch(Dispatchers.Main) {
-                    val intent = Intent(this@MainActivity, MainActivity::class.java)
-                    intent.putExtra("DataAsJson", gson.toJson(data))
-                    startActivity(intent)
+        main_nav.setOnNavigationItemSelectedListener{
+            when(it.itemId){
+                    R.id.left_menu ->{
+                        setFragment(LeftFragment())
+                    }
+                    R.id.middle_menu -> {
+                        setFragment(MiddleFragment())
+                    }
+                    R.id.right_menu -> {
+                        setFragment(RightFragment())
+                    }
+                else -> {
+                    setFragment(MiddleFragment())
                 }
             }
-        }
     }
-
-    private fun alertForInternet() {
-        lateinit var alertDialog: AlertDialog
-        val builder = AlertDialog.Builder(this)
-        builder.setTitle("No Internet!!")
-        builder.setMessage("Turn On Internet!!")
-        builder.setIcon(android.R.drawable.ic_dialog_alert)
-
-        builder.setPositiveButton("Retry") { dialogInterface, which ->
-            if (isNetworkAvailable() == true) {
-                alertDialog.cancel()
-            }
-            else
-            {
-                alertForInternet()
-            }
-        }
-        builder.setNegativeButton("Quit") { dialogInterface, which ->
-            finish();
-            System.exit(0);
-        }
-        alertDialog = builder.create()
-        alertDialog.setCancelable(false)
-        alertDialog.show()
-    }
-
-
-
-    private fun isNetworkAvailable(): Boolean {
-        val cm = getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager
-        val activeNetwork = cm.activeNetworkInfo
-        if (activeNetwork != null) {
-            // connected to the internet
-            if (activeNetwork.type == ConnectivityManager.TYPE_WIFI) {
-                return true
-            } else if (activeNetwork.type == ConnectivityManager.TYPE_MOBILE) {
-                return true
-            }
-            return true
-        }
-        return false
+}
+    private fun setFragment(fragment: Fragment): Boolean {
+        supportFragmentManager.beginTransaction().replace(R.id.main_frame, fragment).commit()
+        return true
     }
 }
