+++ app/src/main/res/menu/bottom_nav.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <item
+        android:id="@+id/left_menu"
+        android:icon="@drawable/left"
+        android:title="1st one" />
+    <item
+        android:id="@+id/middle_menu"
+        android:icon="@drawable/middle"
+        android:title="2nd one" />
+    <item
+        android:id="@+id/right_menu"
+        android:icon="@drawable/right"
+        android:title="3rd one" />
+</menu>
